oh ya its working super fast 4 secs for most generations if there is no extraction that needs to take place

+++ src/org/adligo/xml_io_generator/ExpandClasspathTask.java
@@ -2,6 +2,7 @@ package org.adligo.xml_io_generator;
 
 import java.io.File;
 import java.io.FileInputStream;
+import java.io.FileOutputStream;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Properties;
@@ -68,6 +69,7 @@ public class ExpandClasspathTask extends Task {
 		this.success = success;
 	}
 
+	@SuppressWarnings("deprecation")
 	@Override
 	public void execute() throws BuildException {
 		
@@ -85,12 +87,13 @@ public class ExpandClasspathTask extends Task {
 					"adligo_jse_lib.properties");
 			props.load(fis);
 			fis.close();
+			String classpathFromAnt = project.getProperty(classpath);
 			
 			List<String> classpathEntries = new ArrayList<String>();
-			if (classpath.indexOf(";") == -1) {
-				classpathEntries.add(classpath);
+			if (classpathFromAnt.indexOf(";") == -1) {
+				classpathEntries.add(classpathFromAnt);
 			} else {
-				StringTokenizer tokenizer = new StringTokenizer(classpath, ";");
+				StringTokenizer tokenizer = new StringTokenizer(classpathFromAnt, ";");
 				while (tokenizer.hasMoreElements()) {
 					classpathEntries.add(tokenizer.nextToken());
 				}
@@ -121,31 +124,54 @@ public class ExpandClasspathTask extends Task {
 				String basePackage = genProps.getProperty("basePackage");
 				
 				String pkg = PackageUtils.packageToDir(basePackage);
-				File [] subs = new File(libRoot + File.separator + 
+				String myClasses = libRoot + File.separator + 
 						"xml-io-exp" + File.separator + 
-						pkg).listFiles();
-				
-				for (int i = 0; i < subs.length; i++) {
-					File file = subs[i];
-					if (log.isInfoEnabled()) {
-						log.info("deleting " + file);
+						pkg;
+				if (log.isWarnEnabled()) {
+					log.warn("checking " + myClasses);
+				}
+				File myF = new File(myClasses);
+				if (myF != null) {
+					File [] subs = myF.listFiles();
+					
+					if (subs != null) {
+						if (log.isWarnEnabled()) {
+							log.warn("deleting " + myClasses);
+						}
+						for (int i = 0; i < subs.length; i++) {
+							File file = subs[i];
+							if (log.isInfoEnabled()) {
+								log.info("deleting " + file);
+							}
+							file.delete();
+						}
 					}
-					file.delete();
 				}
 			}
 			for (String jarOrDir: classpathEntries) {
 				int dot = jarOrDir.indexOf(".");
 				if (dot == -1) {
 					//its a dir so recursive copy
-					log.warn("Copy Dir currently not implemented", 
+					log.warn("Copy Dir currently not implemented didn't expand " + jarOrDir, 
 							new Exception("Exception for log importance only, Non Fatal."));
 				} else {
-					int lastSlash = jarOrDir.lastIndexOf(File.separator);
-					String justNamePartOfJar = jarOrDir.substring(
-							lastSlash, jarOrDir.length() - lastSlash - 4);
 					if (log.isInfoEnabled()) {
-						log.info("checking if " + justNamePartOfJar + " has already been expanded.");
+						log.info("checking if expansion is needed for " + jarOrDir);
 					}
+					char [] chars = jarOrDir.toCharArray();
+					StringBuilder sb = new StringBuilder();
+					for (int i = 0; i < chars.length; i++) {
+						char c = chars[i];
+						if (c == File.separatorChar) {
+							sb = new StringBuilder();
+						} else if (c == '.') {
+							break;
+						} else {
+							sb.append(c);
+						}
+					}
+					String justNamePartOfJar = sb.toString();
+					
 					String key = "expanded_jar_" + justNamePartOfJar;
 					String tf = (String) props.get(key);
 					if (!"true".equals(tf)) {
@@ -157,9 +183,15 @@ public class ExpandClasspathTask extends Task {
 							log.warn("to " + libRoot + File.separator + "xml-io-exp");
 						}
 						zu.unzip(in, out);
+						props.setProperty(key, "true");
 					}
 				}
 			}
+			FileOutputStream fos = new FileOutputStream(libRoot + File.separator + 
+					"adligo_jse_lib.properties");
+			props.save(fos, "Expanded several jars which don't need to "
+					+ "be expaneded again until a full jse-main-build.xml");
+			fos.close();
 			project.setUserProperty(success, "true");
 		} catch (Exception x) {
 			project.setUserProperty(success, "false");
